img-recommend/requirements.txt
# -f https://download.pytorch.org/whl/torch_stable.html
# numpy==1.18
# pillow==8.3.2
# facenet_pytorch==2.5.2
# torch==1.8.1
# torchvision==0.9.1

# numpy==1.18.0
# pandas==1.2.4
# pillow==8.3.2
# torch==1.9.0
# facenet_pytorch==2.5.2 -f https://download.pytorch.org/whl/torch_stable.html

-f https://download.pytorch.org/whl/torch_stable.html
numpy==1.19.2
pandas==1.2.4
pillow==7.2.0
torch==1.8.1+cpu
facenet_pytorch==2.5.2
torchvision==0.9.1+cpu
